Convert uiPhotoViewer to a class component
Also swap d3-dispatch for EventEmitter, as this is the more standard way
